20170518
beta release of gaiawatcher.
-stable socket
-stable api
-stable database
-stable web

+++ app_socket/app_socket/GaiaWatcher/Classes/Io.cs
@@ -31,11 +31,11 @@ namespace GaiaWatcher {
             get;
             set;
         }
-        public double batt {
+        public int batt {
             get;
             set;
         }
-        public double vcc {
+        public int vcc {
             get;
             set;
         }
